feat(trusty-common): shared ensure-project-indexed helper; wire tcode task start
Promote trusty-mpm's register-and-populate trusty-search index logic into
`trusty_common::search_index::ensure_project_indexed` (gated behind a new
`search-index` feature that only turns on reqwest's blocking client). trusty-mpm's
`register_project_index` is now a thin wrapper — behaviour unchanged. Expose
`trusty_common::find_git_root` for a cheap git-repo short-circuit. Wire a
best-effort, non-blocking, git-gated ensure-index call into tcode at task START
(both the `run_task` CLI path and the daemon `task::executor` path) so a run's
working project is discoverable via trusty-search while the loop proceeds; the
`index_is_fresh` predicate tests move to trusty-common alongside the logic.

Closes #2700
