+++ crates/trusty-code/src/task/executor.rs
@@ -168,6 +168,14 @@ async fn run_and_record(
     params: TaskRunParams,
     cancel: Arc<AtomicBool>,
 ) {
+    // Trusty-search-first discovery (PR B): at task START, best-effort/detached,
+    // ensure the working project is indexed so the delegated engineer's
+    // `search`/`grep` tools are useful while this run proceeds. Fail-open,
+    // git-gated, and non-blocking — reuses the ONE shared helper (see
+    // `run_task::ensure_project_indexed_in_background`) so the daemon and CLI
+    // paths can never diverge.
+    crate::run_task::ensure_project_indexed_in_background(params.project.clone());
+
     let session_id = params.session_id.clone();
     let transcript: SharedTranscript = Arc::new(Mutex::new(Vec::new()));
     let sink: Arc<dyn ToolEventSink> = Arc::new(SessionToolEventSink::new(
